mit return value erneut test

+++ eBPF3.py
@@ -45,7 +45,7 @@ int inums_update(unsigned int inum) {
 
 static int inums_lookup(unsigned int inum){
     int inum_init();
-    unsigned int *value = inums.lookup(&inum);
+    u64 *value = inums.lookup(&inum);
     if (value==NULL) {
         return 1;  // Wert inum im Array gefunden
     }
